% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_TripPlanner.R
\name{filter_by_season}
\alias{filter_by_season}
\title{Filter Weather Data}
\usage{
filter_by_season(
  location,
  start_date = "2014-01-01",
  end_date = as.character(lubridate::today()),
  season
)
}
\arguments{
\item{location}{A location in the format 'City/zip code, Country'.}

\item{start_date}{The start date for the weather data in the format "YYYY-MM-DD".
The default value is 2014-01-01.}

\item{end_date}{The end date for the weather data in the format "YYYY-MM-DD".
The default value is today's date.}

\item{season}{A season (winter, spring, summer, fall).}
}
\value{
a filtered data frame containing the original data for all of the months
of the indicated season.
}
\description{
filter_by_season filters a data frame by the indicated season.
}
\examples{
filter_by_season("Munich, Germany", "2023-01-10", "2024-05-15", "Spring")
filter_by_season("Leiden, Netherlands", season = "Summer")
filter_by_season("08001, Spain", season = "Winter")
}
