% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_TripPlanner.R
\name{generate_intervals}
\alias{generate_intervals}
\title{Consecutive Day Intervals with Average Weather Parameters}
\usage{
generate_intervals(
  num_days,
  location,
  start_date = "2014-01-01",
  end_date = as.character(lubridate::today()),
  season
)
}
\arguments{
\item{num_days}{The number of consecutive days in each interval.}

\item{location}{A location in the format 'City/zip code, Country'.}

\item{start_date}{The start date for the weather data in the format "YYYY-MM-DD".
The default value is 2014-01-01.}

\item{end_date}{The end date for the weather data in the format "YYYY-MM-DD".
The default value is today's date.}

\item{season}{A season (winter, spring, summer, fall).}
}
\value{
A data frame containing all possible intervals of consecutive days
with the average weather parameters.
}
\description{
generate_intervals creates all possible intervals of
consecutive days based on the specified number of days. The intervals contain
the averages of the weather parameters for all days of the months in the
indicated season over the years.
}
\examples{
generate_intervals(5, "Munich, Germany", "2023-01-10", "2024-05-15", "Spring")
generate_intervals(7, "Leiden, Netherlands", season = "Summer")
generate_intervals(3, "08001, Spain", season = "Winter")
}
