% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_TripPlanner.R
\name{get_weather_data}
\alias{get_weather_data}
\title{Get Weather Data}
\usage{
get_weather_data(
  location,
  start_date = "2014-01-01",
  end_date = as.character(lubridate::today())
)
}
\arguments{
\item{location}{A location in the format 'City/zip code, Country'.}

\item{start_date}{The start date for the weather data in the format "YYYY-MM-DD".
The default value is 2014-01-01}

\item{end_date}{The end date for the weather data in the format "YYYY-MM-DD".
The default value is today's date.}
}
\value{
The parsed weather data as a list.
}
\description{
get_weather_data is used to retrieve historical weather data from the
Open-Meteo API for a specific location and date range.
}
\examples{
get_weather_data("Amsterdam, Netherlands", "2014-01-01", "2014-01-02")
get_weather_data("Amsterdam, Netherlands")
get_weather_data("Tokyo, Japan", "2024-03-03", "2024-03-15")
get_weather_data("Tokyo, Japan")
}
