% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_TripPlanner.R
\name{rank_intervals_precip}
\alias{rank_intervals_precip}
\title{Rank Intervals by Precipitation}
\usage{
rank_intervals_precip(
  num_days,
  location,
  start_date = "2014-01-01",
  end_date = as.character(lubridate::today()),
  season,
  precip_pref
)
}
\arguments{
\item{precip_pref}{The ranking criteria: "none", "snow", or "rain".}

\item{intervals}{A list of data frames containing intervals of consecutive days with weather data.}
}
\value{
A list of ranked intervals based on the precipitation preference.
}
\description{
rank_intervals_temp first calculates the total precipitation,
snowfall, and rain for each interval. It then ranks the intervals based on
the specified precipitation preference ("none", "snow", or "rain")
}
\examples{
rank_intervals_precip(5, "Munich, Germany", "2023-01-10", "2024-05-15", "spring", "none")
rank_intervals_precip(7, "Amsterdam, Netherlands", "2024-01-01", "2024-01-10", "winter", "snow")
rank_intervals_precip(7, "Milan, Italy", "2024-01-01", "2024-03-01", "winter", "rain")
}
