% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_TripPlanner.R
\name{rank_intervals_temp}
\alias{rank_intervals_temp}
\title{Rank Intervals by Temperature}
\usage{
rank_intervals_temp(
  num_days,
  location,
  start_date = "2014-01-01",
  end_date = as.character(lubridate::today()),
  season,
  temp_pref
)
}
\arguments{
\item{temp_pref}{The ranking criteria: "warm" to rank by highest temperature or "cold" to rank by lowest temperature.}

\item{intervals}{A list of data frames containing intervals of consecutive days with weather data.}
}
\value{
A list of ranked intervals based on temperature criteria.
}
\description{
rank_intervals_temp ranks the intervals based on temperature preference.
}
\examples{
rank_intervals_temp(5, "Munich, Germany", "2023-01-10", "2024-05-15", "spring", "warm")
rank_intervals_temp(7, "Amsterdam, Netherlands", "2024-01-01", "2024-01-10", "winter", "cold")
rank_intervals_temp(7, "Milan, Italy", "2024-01-01", "2024-03-01", "winter", "cold")
}
