% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_TripPlanner.R
\name{get_coordinates}
\alias{get_coordinates}
\title{Get Coordinates}
\usage{
get_coordinates(location)
}
\arguments{
\item{location}{A location in the format 'City, Country' (optional with zip code).}
}
\value{
A list with the coordinates (latitude and longitude) of the given location.
The keys of the list are "latitude" and "longitude", and their respective values
are the latitude and longitude coordinates extracted from the API response.
}
\description{
get_coordinates is used to get the latitude and longitude of any location
given as input when calling the function.
}
\examples{
get_coordinates()
# Enter a location (City, Country): Amsterdam, Netherlands
}
