% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_TripPlanner.R
\name{get_weather_data}
\alias{get_weather_data}
\title{Get Weather Data}
\usage{
get_weather_data(
  latitude = latitude,
  longitude = longitude,
  start_date = start_date,
  end_date = end_date
)
}
\arguments{
\item{latitude}{The latitude of the location.}

\item{longitude}{The longitude of the location.}

\item{start_date}{The start date for the weather data in the format "YYYY-MM-DD".}

\item{end_date}{The end date for the weather data in the format "YYYY-MM-DD".}
}
\value{
The parsed weather data as a list.
}
\description{
get_weather_data is used to retrieve historical weather data from the
Open-Meteo API for a specific location and date range.
}
\examples{
get_weather_data()
# Enter a location (City, Country): Linz, Austria
}
