% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_TripPlanner.R
\name{process_weather_data}
\alias{process_weather_data}
\title{Process Weather Data}
\usage{
process_weather_data(weather_data)
}
\arguments{
\item{weather_data}{A list which is returned by the get_weather_data function.}
}
\value{
A data frame with the weather data for each day in the indicated
time period at the gven location.
}
\description{
process_weather_data processes the data which was retrieved with the
get_weather_data function and creates a new data frame with the daily weather data.
}
\examples{
process_weather_data(weather_data)
}
